obfuscate old forge config guis - saves ~500 bytes just need the providers also not tested
